+++ RazorSvelte/App/shared/config.ts
@@ -21,6 +21,7 @@ export let user = getValueFromJson<{isSigned: boolean, name?: string, email?: st
 
 export const errorKey = getValue<string>("error-key");
 export const themeKey = getValue<string>("theme-key");
+export const title = getValue<string>("title");
 export const urls = getValueFromJson<{
     authorizedUrl: string;
     indexUrl: string;