+++ RazorSvelte/Pages/_Consts.cs
@@ -5,4 +5,5 @@ public class Consts
     public const string ThemeKey = "__theme";
     public const string ErrorKey = "__error";
     public const string ApiSegment = "/api";
+    public const string Title = "RazorSvelte";
 }